+++ packages/pure-cell/src/components/footer-text/component.tsx
@@ -1,9 +1,9 @@
-import React, { ReactNode, useContext } from 'react';
+import React, { type ReactNode, useContext } from 'react';
 
 import { getDataTestId } from '@alfalab/core-components-shared';
 import { type Color, Text } from '@alfalab/core-components-typography';
 
-import { PureCellContext } from '../../component';
+import { PureCellContext } from '../../context';
 
 export type FooterTextProps = {
     /**