+++ packages/toast-plate/src/shared/index.ts
@@ -1,0 +1,1 @@
+export { getBaseToastPlateTestIds } from '../components/base-toast-plate/utils/getBaseToastPlateTestIds';