+++ packages/system-message/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-system-message
 
+## 3.0.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-shared@1.1.0
+    -   @alfalab/core-components-mq@5.0.1
+
 ## 3.0.0
 
 ### Major Changes